This CL adds a few hacks to libgcc_eh to make it work for x86-64:
* it adds a call to the new "intrinsic" pnacl_unwind_init_dwarf_reg_size_table

* it asserts the settings of various global #defines
because they influence the code path.
This is mostly meant to document what assumptions our bitcode libgcc_eh makes.
As long as we are building libgcc_eh natively, this also ensures that those settings hold for all
three currently supported archs.

* It disables a few asserts that fire for x86-64.
Those asserts are justified and we lose the upper 32bits when certain 64 bit registers are restored.
Luckily those regs are typically pointers where this does not matter. This needs to be carefully revisited in a future CL. It is not clear that native version of libgcc_eh gets this right either
because in the x86-64 case the frontend is assuming sizeof(void*) == 8 but when we compile the bitcode to native code we actually assume sizeof(void*) == 4
c.f. http://code.google.com/p/nativeclient/issues/detail?id=869

* Improved the debugging helpers somewhat. Note: The helpers are "defined out" by default

BUG= http://code.google.com/p/nativeclient/issues/detail?id=869
BUG= http://code.google.com/p/nativeclient/issues/detail?id=1520

CL= http://codereview.chromium.org/7054087/